requirements.txt
black==23.11.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
docker==7.1.0
dotenv==0.9.9
flake8==6.1.0
idna==3.10
iniconfig==2.1.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.19.1
pytest==8.0.2
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
urllib3==2.3.0
